fix: fixes linking problems related to xByte2String method
