+++ src/arduino_helpers.c
@@ -1,0 +1,28 @@
+#include "arduino_helpers.h"
+
+#ifdef __cplusplus
+
+#if defined(ARDUINO_ARCH_AVR) || defined(ARDUINO_ARCH_SAM) || defined(ARDUINO_ARCH_SAMD) || defined(ARDUINO_ARCH_ESP8266) || defined(ARDUINO_ARCH_STM32) || \
+    defined(ARDUINO_TEENSY_MICROMOD) || defined(ARDUINO_TEENSY40) || defined(ARDUINO_TEENSY41) || defined(ARDUINO_TEENSY36) || defined(ARDUINO_TEENSY35) || \
+    defined(ARDUINO_TEENSY31) || defined(ARDUINO_TEENSY32) || defined(ARDUINO_TEENSY30) || defined(ARDUINO_TEENSYLC)
+
+String xByte2String(xSize size_, xByte *bytes_) {
+    String str = "";
+    xSize i = 0;
+    char buf[size_ + 1];
+
+    if(NOTNULLPTR(bytes_) && (zero < size_)) {
+        for(i = 0; i < size_; i++) {
+            buf[i] = (char) bytes_[i];
+        }
+
+        buf[size_] = '\0';
+        str = String(buf);
+    }
+
+    return(str);
+}
+
+#endif
+
+#endif /* ifdef __cplusplus */
